+++ src/app/mockdata/inspection-question-data.ts
@@ -1,0 +1,3 @@
+export class InspectionQuestionData {
+  public data: any[] = [];
+}
