+++ src/app/mockdata/survey-question-data.ts
@@ -1,0 +1,299 @@
+export class SurveyQuestionData {
+  public data: any[] = [{
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '432997bb-7997-4450-9a61-ec02c14f7ca5',
+    'isActive': true,
+    'description': {'idLanguageContent': '3bb62696-920a-4b00-bc40-449564ea14de', 'fr': 'Ajouter un nouveau détecteur'},
+    'idLanguageContentTitle': '70f113df-ed6c-4da9-baea-5b95064b99e9',
+    'idSurveyQuestionNext': null,
+    'idSurveyQuestionNextOnFalse': '6d5da791-ddcd-4315-a472-bec6817f10a0',
+    'title': {'idLanguageContent': '70f113df-ed6c-4da9-baea-5b95064b99e9', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '3bb62696-920a-4b00-bc40-449564ea14de',
+    'questionType': 'boolean',
+    'sequence': 1
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '875943f9-9218-4818-8e3f-ef7e65aa8615',
+    'isActive': true,
+    'description': {'idLanguageContent': '8a8fd6cd-35ba-4d51-8567-a73a6dca49e3', 'fr': 'Est-il fonctionnel ?'},
+    'idLanguageContentTitle': '0d0bab2c-30b8-4fe2-9f8a-ef60803623c3',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '0d0bab2c-30b8-4fe2-9f8a-ef60803623c3', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '8a8fd6cd-35ba-4d51-8567-a73a6dca49e3',
+    'questionType': 'boolean',
+    'sequence': 2
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '75a256f9-2d6f-457d-8d7d-7899b3238e32',
+    'isActive': true,
+    'description': {'idLanguageContent': '73b58356-8511-4307-911c-c6e6cfbe5650', 'fr': 'L\'emplacement est-il approprié ?'},
+    'idLanguageContentTitle': '870c3d26-9fc9-4aa6-814c-72b5806f24da',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '870c3d26-9fc9-4aa6-814c-72b5806f24da', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '73b58356-8511-4307-911c-c6e6cfbe5650',
+    'questionType': 'boolean',
+    'sequence': 3
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '04e97b3a-97cf-4c0c-af1b-3f76ee8a4539',
+    'isActive': true,
+    'description': {'idLanguageContent': '690a7e2a-e971-400d-9cd0-bda4b234676b', 'fr': 'Est-il alimenté avec une pile ?'},
+    'idLanguageContentTitle': '19a8c8c3-0d1d-4922-9a24-0e345ed16179',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '19a8c8c3-0d1d-4922-9a24-0e345ed16179', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '690a7e2a-e971-400d-9cd0-bda4b234676b',
+    'questionType': 'boolean',
+    'sequence': 4
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': 'a41a3998-2b53-4678-9b5f-2c481087885c',
+    'isActive': true,
+    'description': {'idLanguageContent': '3c58faf5-519c-4a88-9f70-38c2a85cc5fa', 'fr': 'Est-il relié au réseau électrique du bâtiment ?'},
+    'idLanguageContentTitle': 'f4b73cae-f695-48bd-a96c-7a2abdb759d7',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': 'f4b73cae-f695-48bd-a96c-7a2abdb759d7', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '3c58faf5-519c-4a88-9f70-38c2a85cc5fa',
+    'questionType': 'boolean',
+    'sequence': 5
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': 'cb0d43b9-d761-46d0-8f5a-c733ac4a4e9d',
+    'isActive': true,
+    'description': {'idLanguageContent': 'af4f8104-836f-4373-ad23-c44f0567de67', 'fr': 'Est-il relié à une centrale ?'},
+    'idLanguageContentTitle': 'ae26b6e2-f6fd-47e8-8340-ad79687fadc6',
+    'idSurveyQuestionNext': '432997bb-7997-4450-9a61-ec02c14f7ca5',
+    'title': {'idLanguageContent': 'ae26b6e2-f6fd-47e8-8340-ad79687fadc6', 'fr': 'Avertisseur de fumée'},
+    'idLanguageContentDescription': 'af4f8104-836f-4373-ad23-c44f0567de67',
+    'questionType': 'boolean',
+    'sequence': 6
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '6d5da791-ddcd-4315-a472-bec6817f10a0',
+    'isActive': true,
+    'description': {
+      'idLanguageContent': '9d38e728-d158-470e-9823-ea2ff0be1779',
+      'fr': 'Autres notes ou commentaires pertinent à écrire au dossier ?'
+    },
+    'idLanguageContentTitle': 'f360055b-ce21-4587-8cdf-e496de4146d6',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': 'f360055b-ce21-4587-8cdf-e496de4146d6', 'fr': 'Avertisseurs de fumée'},
+    'idLanguageContentDescription': '9d38e728-d158-470e-9823-ea2ff0be1779',
+    'questionType': 'text',
+    'sequence': 7
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '432997bb-7997-4450-9a61-ec02c14f7ca4',
+    'isActive': true,
+    'description': {'idLanguageContent': '3bb62696-920a-4b00-bc40-449564ea14d4', 'fr': 'Ajouter un nouveau type de chauffage'},
+    'idLanguageContentTitle': '70f113df-ed6c-4da9-baea-5b95064b99e4',
+    'idSurveyQuestionNext': null,
+    'idSurveyQuestionNextOnFalse': '59ee70db-5372-4ba3-9adf-343283d3124f',
+    'title': {'idLanguageContent': '70f113df-ed6c-4da9-baea-5b95064b99e4', 'fr': 'Chauffage'},
+    'idLanguageContentDescription': '3bb62696-920a-4b00-bc40-449564ea14d4',
+    'questionType': 'boolean',
+    'sequence': 8
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '79ec6e47-f267-46de-a511-510a7567c670',
+    'isActive': true,
+    'description': {'idLanguageContent': 'db181fce-f930-4a23-ae37-eba75f7af4ed', 'fr': 'De quel type de chauffage s\'agit-il ?'},
+    'idLanguageContentTitle': '0008556a-2a4d-448a-bf43-ded4dca5b35d',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '0008556a-2a4d-448a-bf43-ded4dca5b35d', 'fr': 'Chauffage'},
+    'idLanguageContentDescription': 'db181fce-f930-4a23-ae37-eba75f7af4ed',
+    'questionType': 'choice',
+    'choices': [{
+      'idSurveyChoice': 'a77aa2fd-2c0b-4611-b3f4-b7a599dc9fcf',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'Électrique',
+        'idLanguageContent': '535dd217-c8a6-417c-908d-7e5c2f346b8c'
+      },
+      'idSurveyQuestionNext': null,
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': '535dd217-c8a6-417c-908d-7e5c2f346b8c'
+    }, {
+      'idSurveyChoice': 'd5689b4f-7b33-44ae-8514-7da0d079f2d8',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'Bois',
+        'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aaa'
+      },
+      'idSurveyQuestionNext': '4bb15adf-2d57-4695-b462-70763fbec212',
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aaa'
+    }, {
+      'idSurveyChoice': 'd5689b4f-7b33-44ae-8514-7da0d079f2d1',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'Huile',
+        'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aa1'
+      },
+      'idSurveyQuestionNext': null,
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aa1'
+    }, {
+      'idSurveyChoice': 'd5689b4f-7b33-44ae-8514-7da0d079f2db',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'Gaz',
+        'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aab'
+      },
+      'idSurveyQuestionNext': null,
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aab'
+    }, {
+      'idSurveyChoice': 'd5689b4f-7b33-44ae-8514-7da0d079f2dc',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'BiÉnergie',
+        'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aac'
+      },
+      'idSurveyQuestionNext': null,
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aac'
+    }, {
+      'idSurveyChoice': 'd5689b4f-7b33-44ae-8514-7da0d079f2dd',
+      'idSurveyQuestion': 'f0ab84f0-e216-49c4-a714-c31420c68abf',
+      'description': {
+        'fr': 'Autre',
+        'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aad'
+      },
+      'idSurveyQuestionNext': null,
+      'createdOn': '2017/04/06 16:19:44',
+      'isActive': true,
+      'sequence': 0,
+      'idLanguageContent': 'c0355f3c-927b-406e-aaa5-0562d1ae3aae'
+    }],
+    'sequence': 9
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '4bb15adf-2d57-4695-b462-70763fbec212',
+    'isActive': true,
+    'description': {'idLanguageContent': '5a61152d-1576-463c-a416-971a5df9e9b2', 'fr': 'Date du dernier ramonage ?'},
+    'idLanguageContentTitle': '9b72bc58-6181-461b-a737-e3fc38ca7292',
+    'idSurveyQuestionNext': '',
+    'title': {'idLanguageContent': '9b72bc58-6181-461b-a737-e3fc38ca7292', 'fr': 'Chauffage - ramonage'},
+    'idLanguageContentDescription': '5a61152d-1576-463c-a416-971a5df9e9b2',
+    'questionType': 'date',
+    'sequence': 10
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '59ee70db-5372-4ba3-9adf-343283d3124f',
+    'isActive': true,
+    'description': {
+      'idLanguageContent': '1b552a81-985f-40ba-8960-16696bd6889d',
+      'fr': 'Il y a-t-il présence d\'un extincteur portatif ?'
+    },
+    'idLanguageContentTitle': '669c32a7-d41d-4a6c-82cf-0978cdaf0424',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '669c32a7-d41d-4a6c-82cf-0978cdaf0424', 'fr': 'Extincteurs portatifs'},
+    'idLanguageContentDescription': '1b552a81-985f-40ba-8960-16696bd6889d',
+    'questionType': 'boolean',
+    'sequence': 11
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': 'af21c292-d913-484c-8d02-cb60ef3ca19f',
+    'isActive': true,
+    'description': {'idLanguageContent': '9cdc1f02-f1cb-4c47-a236-7d0255547c0a', 'fr': 'Est-il entretenu et vérifié régulièrement ?'},
+    'idLanguageContentTitle': '6817d986-6d6e-492a-8e58-c6c6311ab1cd',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '6817d986-6d6e-492a-8e58-c6c6311ab1cd', 'fr': 'Extincteurs portatifs'},
+    'idLanguageContentDescription': '9cdc1f02-f1cb-4c47-a236-7d0255547c0a',
+    'questionType': 'boolean',
+    'sequence': 12
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': 'daef1dba-89f0-41f4-99e3-1cd7d8798349',
+    'isActive': true,
+    'description': {'idLanguageContent': '02f1bebf-3264-46bd-9684-0ca33b6dd2dc', 'fr': 'Il y a-t-il présence d\'un avertisseur de CO ?'},
+    'idLanguageContentTitle': '496864c0-1cf1-4988-86c8-581d08834713',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '496864c0-1cf1-4988-86c8-581d08834713', 'fr': 'Avertisseurs de monoxyde de carbone'},
+    'idLanguageContentDescription': '02f1bebf-3264-46bd-9684-0ca33b6dd2dc',
+    'questionType': 'boolean',
+    'sequence': 13
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '562b737a-2c82-4d05-a283-db7dd684ed72',
+    'isActive': true,
+    'description': {'idLanguageContent': '3566e586-2fd0-4791-adf9-c04795e373e6', 'fr': 'L\'emplacement est-il approprié ?'},
+    'idLanguageContentTitle': 'c3537964-a5db-4751-a640-4ec1b623ec73',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': 'c3537964-a5db-4751-a640-4ec1b623ec73', 'fr': 'Avertisseurs de monoxyde de carbone'},
+    'idLanguageContentDescription': '3566e586-2fd0-4791-adf9-c04795e373e6',
+    'questionType': 'boolean',
+    'sequence': 14
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '6f5eecc6-a08e-46eb-9952-aa3b4c7c76f4',
+    'isActive': true,
+    'description': {'idLanguageContent': '0f16ac00-871f-4372-8142-a9784ce87da2', 'fr': 'Adresse visible de la voie publique ?'},
+    'idLanguageContentTitle': '58d2c928-1c30-4fc7-8403-1ba212f55e51',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '58d2c928-1c30-4fc7-8403-1ba212f55e51', 'fr': 'Visibilité'},
+    'idLanguageContentDescription': '0f16ac00-871f-4372-8142-a9784ce87da2',
+    'questionType': 'boolean',
+    'sequence': 15
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '76a21a3a-58bd-4aef-9ba3-68d179b26ad9',
+    'isActive': true,
+    'description': {'idLanguageContent': '20acad3d-ff5d-47d0-870b-b9f314d1ab8f', 'fr': 'Présence d\'une personne à mobilité réduite ?'},
+    'idLanguageContentTitle': '8f2784ae-2a6a-4866-841a-556222efe37d',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '8f2784ae-2a6a-4866-841a-556222efe37d', 'fr': 'PNAP'},
+    'idLanguageContentDescription': '20acad3d-ff5d-47d0-870b-b9f314d1ab8f',
+    'questionType': 'text',
+    'sequence': 16
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '6da3d453-819a-448c-a025-8bdfbe73fabd',
+    'isActive': true,
+    'description': {
+      'idLanguageContent': '2d5d520f-0590-4520-87cb-1ccfa6fb1f23',
+      'fr': 'Est-ce que le dossier nécessite un transfert à la prévention ?'
+    },
+    'idLanguageContentTitle': 'c12b4b9e-f34d-4f53-ba01-18cbee357c91',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': 'c12b4b9e-f34d-4f53-ba01-18cbee357c91', 'fr': 'Transfert'},
+    'idLanguageContentDescription': '2d5d520f-0590-4520-87cb-1ccfa6fb1f23',
+    'questionType': 'boolean',
+    'sequence': 17
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '09332791-3e96-4b1e-8aa4-237599fe53f1',
+    'isActive': true,
+    'description': {'idLanguageContent': 'fd4eefea-c763-4c85-888d-222b58d2c70d', 'fr': 'Quelle est la raison ?'},
+    'idLanguageContentTitle': '02255afb-b14a-4542-89a8-210336df9e6b',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '02255afb-b14a-4542-89a8-210336df9e6b', 'fr': 'Divers'},
+    'idLanguageContentDescription': 'fd4eefea-c763-4c85-888d-222b58d2c70d',
+    'questionType': 'text',
+    'sequence': 18
+  }, {
+    'idSurvey': '43deacf8-fef0-4b88-a0fd-29cb1bfa0e04',
+    'idSurveyQuestion': '89f41075-33e2-4139-a0b5-0adce2a23342',
+    'isActive': true,
+    'description': {
+      'idLanguageContent': 'c3aceceb-4500-4c60-868c-976a8ffb460d',
+      'fr': 'Autres notes ou commentaires pertinent à écrire au dossier ?'
+    },
+    'idLanguageContentTitle': '594af314-a543-46f8-b1b5-9f9a4789b00b',
+    'idSurveyQuestionNext': null,
+    'title': {'idLanguageContent': '594af314-a543-46f8-b1b5-9f9a4789b00b', 'fr': 'Notes'},
+    'idLanguageContentDescription': 'c3aceceb-4500-4c60-868c-976a8ffb460d',
+    'questionType': 'text',
+    'sequence': 19
+  }];
+}
