+++ src/app/prevention-survey/prevention-survey.module.ts
@@ -3,7 +3,8 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
 import { QuestionComponent } from './question/question.component';
 import { QuestionSliderComponent } from './question-slider/question-slider.component';
-import { QuestionService } from './shared/services/question.service';
+import { SurveyQuestionService } from './shared/services/survey-question.service';
+import { InspectionQuestionService } from './shared/services/inspection-question.service';
 
 @NgModule({
   imports: [
@@ -17,7 +18,8 @@ import { QuestionService } from './shared/services/question.service';
     QuestionSliderComponent,
   ],
   providers: [
-    QuestionService
+    SurveyQuestionService,
+    InspectionQuestionService
   ]
 })
 export class PreventionSurveyModule { }
