+++ src/app/prevention-survey/question/question.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import {Component, EventEmitter, Input, Output, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-prevention-survey-question',
@@ -6,11 +6,27 @@ import { Component, Input, OnInit } from '@angular/core';
   styleUrls: ['./question.component.styl']
 })
 export class QuestionComponent implements OnInit {
-  @Input() question: Object;
+  @Input() question: object;
+  @Output() change: EventEmitter<boolean|string> = new EventEmitter<boolean|string>();
 
   constructor() { }
 
   ngOnInit() {
   }
 
+  boolChange(e) {
+    this.change.emit(e.checked);
+  }
+
+  choiceChange(e) {
+    this.change.emit(e);
+  }
+
+  dateChange(e) {
+    this.change.emit(e.target.value);
+  }
+
+  textChange(e) {
+    this.change.emit(e.target.value);
+  }
 }
