+++ src/app/prevention-survey/shared/services/inspection-question.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class InspectionQuestionService {
+  private answers: object[];
+
+  constructor() { }
+
+  answer(idSurveyQuestion: string, answer: any) {
+    this.answers.push({
+      idSurveyQuestion: answer
+    });
+  }
+}
