+++ src/app/prevention-survey/shared/services/survey-question.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { SurveyQuestionService } from './survey-question.service';
+
+describe('SurveyQuestionService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [SurveyQuestionService]
+    });
+  });
+
+  it('should ...', inject([SurveyQuestionService], (service: SurveyQuestionService) => {
+    expect(service).toBeTruthy();
+  }));
+});
